Splitting branch into own component
