+++ frontend/src/views/AdminPages/BranchesPage/BranchesPage.js
@@ -1,13 +1,8 @@
 import React, { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { Link } from "react-router-dom";
 import { listBranches, saveBranch, deleteBranch } from '../../../actions/branchesActions';
 // @material-ui/core components
 import { makeStyles } from "@material-ui/core/styles";
-import Checkbox from "@material-ui/core/Checkbox";
-import FormControlLabel from "@material-ui/core/FormControlLabel";
-// @material-ui/icons
-import Check from "@material-ui/icons/Check";
 // core components
 import Header from "components/Header/Header.js";
 import HeaderLinks from "components/Header/HeaderLinks.js";
@@ -18,7 +13,6 @@ import Card from "components/Card/Card.js";
 import CardBody from "components/Card/CardBody.js";
 import CardHeader from "components/Card/CardHeader.js";
 import CardFooter from "components/Card/CardFooter.js";
-import CustomInput from "components/CustomInput/CustomInput";
 import Button from "components/CustomButtons/Button.js";
 import SneakingComponent from "../SneakingComponent/SneakingComponent"
 // assets
@@ -27,6 +21,7 @@ import checkboxstyles from "assets/jss/material-kit-react/customCheckboxRadioSwi
 import bgimage from "assets/img/bg7.jpg";
 import Danger from "components/Typography/Danger";
 import BranchComponent from "./BranchComponent/BranchComponent";
+import BranchEditComponent from "./BranchEditComponent/BranchEditComponent";
 
 const useStyles = makeStyles(styles);
 const cbStyles = makeStyles(checkboxstyles);
@@ -37,13 +32,7 @@ export default function BranchesPage(props) {
     const { ...rest } = props;
 
     const [modalVisible, setModalVisible] = useState(false);
-    const [id, setId] = useState('');
-    const [city, setCity] = useState('');
-    const [description, setDescription] = useState('');
-    const [image, setImage] = useState('');
-    const [longitude, setLongitude] = useState('');
-    const [latitude, setLatitude] = useState('');
-    
+    const [branch, setBranch] = useState();
 
     const userSignin = useSelector(state => state.userSignin);
     const { userInfo } = userSignin
@@ -67,24 +56,10 @@ export default function BranchesPage(props) {
 
     const closeModal = () => {
         setModalVisible(false);
-        setId(''); // if we discard an edited branch and create a new one, we get old fields but new id
     }
     const openModal = (branch) => {
+        setBranch(branch);
         setModalVisible(true);
-        setId(branch._id);
-        setCity(branch.branchCity);
-        setDescription(branch.branchDescription);
-        setImage(branch.branchImage);
-        setLongitude(branch.branchLocation && branch.branchLocation.coordinates[0]);
-        setLatitude(branch.branchLocation && branch.branchLocation.coordinates[1]);
-    }
-
-    const submitHandler = (e) => {
-        e.preventDefault();
-        dispatch(saveBranch({
-            _id: id,
-            city, description, image, longitude, latitude
-        }))
     }
 
     const deleteHandler = (branch) => {
@@ -152,111 +127,18 @@ export default function BranchesPage(props) {
                                         </table>
                                     </CardBody>
                                     <CardFooter className={classes.cardFooter}>
-                                        <Button type="button" simple size="lg" color="primary" onClick={() => openModal({})}>New Branch</Button>
+                                        <Button type="button" simple size="lg" color="primary" onClick={() => openModal()} disabled={modalVisible}>New Branch</Button>
                                     </CardFooter>
                                 </form>
                             </Card>
 
                             {/* Create or update branch card */}
                             {modalVisible &&
-                                <GridContainer justify="center">
-                                    <GridItem xs={12} sm={12} md={12}>
-                                        <Card>
-                                            <form onSubmit={submitHandler} className={classes.form}>
-                                                <GridContainer justify="center">
-                                                    <GridItem xs={12} sm={12} md={4}>
-                                                        <CardHeader color="primary" className={classes.cardHeader}>
-                                                            <h4>Branch Details</h4>
-                                                        </CardHeader>
-                                                    </GridItem>
-                                                </GridContainer>
-                                                {loadingSave && <p className={classes.divider}>Loading...</p>}
-                                                {errorSave && <p className={classes.divider}><Danger>Error saving info. Please validate the fields and try again later</Danger></p>}
-                                                <CardBody>
-                                                    <GridContainer>
-                                                        <GridItem xs={12} sm={6} md={4}>
-                                                            <CustomInput
-                                                                labelText="Branch City"
-                                                                id="city"
-                                                                formControlProps={{
-                                                                    fullWidth: true,
-                                                                }}
-                                                                inputProps={{
-                                                                    type: "text",
-                                                                    value: city,
-                                                                    onChange: (e) => setCity(e.target.value),
-                                                                }}
-                                                            />
-                                                        </GridItem>
-                                                        <GridItem xs={12} sm={6} md={4}>
-                                                            <CustomInput
-                                                                labelText="Branch Description"
-                                                                id="description"
-                                                                formControlProps={{
-                                                                    fullWidth: true,
-                                                                }}
-                                                                inputProps={{
-                                                                    type: "text",
-                                                                    value: description,
-                                                                    onChange: (e) => setDescription(e.target.value)
-                                                                }}
-                                                            />
-                                                        </GridItem>
-                                                        <GridItem xs={12} sm={12} md={4}>
-                                                            <CustomInput
-                                                                labelText="Image"
-                                                                id="image"
-                                                                formControlProps={{
-                                                                    fullWidth: true,
-                                                                }}
-                                                                inputProps={{
-                                                                    type: "text",
-                                                                    value: image,
-                                                                    onChange: (e) => setImage(e.target.value)
-                                                                }}
-                                                            />
-                                                        </GridItem>
-                                                        <GridItem xs={12} sm={12} md={4}>
-                                                            <CustomInput
-                                                                labelText="Longitude (E/W)"
-                                                                id="longitude"
-                                                                formControlProps={{
-                                                                    fullWidth: true,
-                                                                }}
-                                                                inputProps={{
-                                                                    type: "number",
-                                                                    value: longitude,
-                                                                    onChange: (e) => setLongitude(e.target.value)
-                                                                }}
-                                                            />
-                                                        </GridItem>
-                                                        <GridItem xs={12} sm={12} md={4}>
-                                                            <CustomInput
-                                                                labelText="Latitude (N/S)"
-                                                                id="latitude"
-                                                                formControlProps={{
-                                                                    fullWidth: true,
-                                                                }}
-                                                                inputProps={{
-                                                                    type: "number",
-                                                                    value: latitude,
-                                                                    onChange: (e) => setLatitude(e.target.value)
-                                                                }}
-                                                            />
-                                                        </GridItem>
-
-                                                    </GridContainer>
-                                                </CardBody>
-                                                <CardFooter className={classes.cardFooter}>
-                                                    <Button simple color="warning" size="lg" onClick={() => closeModal()}>
-                                                        Discard
-                                                    </Button>
-                                                    <Button simple color="primary" size="lg" type="submit">{id ? "Update" : "Create"} Branch</Button>
-                                                </CardFooter>
-                                            </form>
-                                        </Card>
-                                    </GridItem>
-                                </GridContainer>
+                                <BranchEditComponent
+                                    branch={branch}
+                                    loadingSave={loadingSave}
+                                    errorSave={errorSave}
+                                    onCloseClick={() => closeModal()} />
                             }
                         </div>
                         <Footer whiteFont />
